+++ lib/yak_merchant_app_design_lib.dart
@@ -10,6 +10,9 @@ export 'inputs/yak_text_area.dart';
 export 'inputs/yak_select.dart';
 export 'inputs/yak_otp_input.dart';
 
+// Indicators
+export 'indicators/yak_indicator.dart';
+
 // Theme
 export 'theme/yak_color.dart';
 export 'theme/yak_typography.dart';